+++ src/url/email.js
@@ -17,6 +17,9 @@ export class Email extends URL {
   }
 
   /**
+   * The email address (`username@domain.tld`) as a string.
+   *
+   * (maintainer comment)
    * Replace the string representation of the top-level class (`URL`) to be an
    * email address instead of `ftp://username@domain.tld`. It is needed for
    * situation where type casting to string is involved (`console.log`…).
@@ -38,7 +41,7 @@ export class Email extends URL {
   /**
    * Whether or not an email address is parsable and valid.
    *
-   * @param {any|Email} address
+   * @param {any|string|Email|Stringable} address
    */
   static canParse(address) {
     try {
@@ -49,3 +52,17 @@ export class Email extends URL {
     }
   }
 }
+
+/**
+ * Whether or not an email address is parsable and valid.
+ *
+ * It uses WHATWG recommended RegExp: https://html.spec.whatwg.org/multipage/input.html#valid-e-mail-address
+ *
+ * @param {any|string|Stringable} address
+ */
+export const isEmail = address => /^[a-zA-Z0-9.!#$%&'*+/=?^_`{|}~-]+@[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?(?:\.[a-zA-Z0-9](?:[a-zA-Z0-9-]{0,61}[a-zA-Z0-9])?)*$/.test(address)
+
+/**
+ * @typedef {Object} Stringable
+ * @property {function(): string} toString - The object as a string.
+ */