+++ CHANGELOG.md
@@ -10,6 +10,14 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 Compare with [last published version](https://github.com/frontacles/frontacles/compare/0.2.3...main).
 
+### New
+
+- Add [`isEmail`](./README.md#isemail) to validate emails.
+
+### Documentation
+
+- It is now more explicit in types that `Email.canParse` expects a `string` or a `Stringable` (from `any|Email` to `any|string|Email|Stringable`).
+
 ### Under the hood
 
 - Centralize all benchmarks in [`./benchs`](./benchs)