+++ CHANGELOG.md
@@ -8,6 +8,8 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 <!-- ⚠️ Before a new release, make sure the documentation doesn't contain any **unreleased** mention.  -->
 
+Compare with [last published version](https://github.com/frontacles/frontacles/compare/0.4.0...main). 
+
 ### New
 
 - Add [`setAttributes`](./README.md#setattributes), a function to update HTML attributes on any number of elements.
@@ -17,8 +19,6 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 - Setup tests for DOM utilities using Playwright via Vitest browser mode.
 - Migrate types testing from `tsd` to TSTyche.
 - Bump Node version from 20 to 22.
-
-Compare with [last published version](https://github.com/frontacles/frontacles/compare/0.4.0...main). 
 
 ## v0.4.0 (2025-03-08)